Simplify instructions for Heimdall installs
It is unnecessary to disconnect USB cable on my gta4xlwifi. Doing so is confusing, as the stock firmware immediately undoes the flashed recovery. Instead, I followed https://forum.xda-developers.com/t/how-to-flash-recovery-using-linux-and-heimdall-solved.4238595/ which suggested holding down the recovery buttons while the flashing is happening.

It's also confusing to tell heimdall to use --no-reboot on the flash, as it gives no indication of being done.